requirements.txt
django
djangorestframework
pdfplumber
spacy==3.7.2
Pillow
pytesseract
opencv-python
spellchecker